Move anidb related utilities to anidbUtils.ts
